+++ govtool/backend/src/VVA/Types.hs
@@ -24,6 +24,7 @@ import           Database.PostgreSQL.Simple (Connection)
 import           VVA.Cache
 import           VVA.Config
 import           Network.HTTP.Client (Manager)
+import           Control.Concurrent.QSem
 
 type App m = (MonadReader AppEnv m, MonadIO m, MonadFail m, MonadError AppError m)
 
@@ -33,6 +34,7 @@ data AppEnv
       , vvaCache          :: CacheEnv
       , vvaConnectionPool :: Pool Connection
       , vvaTlsManager     :: Manager
+      , vvaMetadataQSem   :: QSem
       }
 
 instance Has VVAConfig AppEnv where
@@ -50,6 +52,11 @@ instance Has (Pool Connection) AppEnv where
 instance Has Manager AppEnv where
   getter AppEnv {vvaTlsManager} = vvaTlsManager
   modifier f a@AppEnv {vvaTlsManager} = a {vvaTlsManager = f vvaTlsManager}
+
+instance Has QSem AppEnv where
+  getter AppEnv {vvaMetadataQSem} = vvaMetadataQSem
+  modifier f a@AppEnv {vvaMetadataQSem} = a {vvaMetadataQSem = f vvaMetadataQSem}
+
 
 data AppError
   = ValidationError Text
@@ -133,6 +140,34 @@ data Proposal
 
 data TransactionStatus = TransactionConfirmed | TransactionUnconfirmed
 
+
+data ProposalMetadata =
+    ProposalMetadata
+    { proposalMetadataAbstract :: Text
+    , proposalMetadataMotivation :: Text
+    , proposalMetadataRationale :: Text
+    , proposalMetadataTitle :: Text
+    , proposalMetadataReferences :: [Text]
+    } deriving (Show)
+
+data DRepMetadata =
+    DRepMetadata
+    { dRepMetadataBio :: Text
+    , dRepMetadataDRepName :: Text
+    , dRepMetadataEmail :: Text
+    , dRepMetadataReferences :: [Text]
+    } deriving (Show)
+
+data MetadataValidationResult a =
+    MetadataValidationResult
+    { metadataValidationResultValid :: Bool
+    , metadataValidationResultStatus :: Maybe Text
+    , metadataValidationResultMetadata :: Maybe a
+    } deriving (Show)
+
+
+
+
 data CacheEnv
   = CacheEnv
       { proposalListCache                  :: Cache.Cache () [Proposal]
@@ -145,7 +180,8 @@ data CacheEnv
       , dRepVotingPowerCache               :: Cache.Cache Text Integer
       , dRepListCache                      :: Cache.Cache () [DRepRegistration]
       , networkMetricsCache                :: Cache.Cache () NetworkMetrics
-      , metadataValidationCache            :: Cache.Cache (Text, Text) Value
+      , proposalMetadataValidationCache            :: Cache.Cache (Text, Text) (MetadataValidationResult ProposalMetadata)
+      , dRepMetadataValidationCache                :: Cache.Cache (Text, Text) (MetadataValidationResult DRepMetadata)
       }
 
 data NetworkMetrics
@@ -175,3 +211,5 @@ data MetadataValidationStatus
   | IncorrectJSONLD
   | IncorrectHash
   | UrlNotFound
+
+