+++ govtool/frontend/src/components/molecules/GovernanceVotedOnCard.tsx
@@ -4,7 +4,6 @@ import { Box } from "@mui/material";
 import { Button } from "@atoms";
 import { PATHS } from "@consts";
 import { useScreenDimension, useTranslation } from "@hooks";
-import { VotedProposalToDisplay } from "@models";
 import {
   formatDisplayDate,
   getFullGovActionId,
@@ -18,9 +17,10 @@ import {
   GovernanceActionCardStatePill,
   GovernanceActionsDatesBox,
 } from "@molecules";
+import { VotedProposal } from "@/models";
 
 type Props = {
-  votedProposal: VotedProposalToDisplay;
+  votedProposal: VotedProposal;
   inProgress?: boolean;
 };
 
@@ -28,16 +28,17 @@ export const GovernanceVotedOnCard = ({ votedProposal, inProgress }: Props) => {
   const navigate = useNavigate();
   const { proposal, vote } = votedProposal;
   const {
-    about,
+    abstract,
     createdDate,
     createdEpochNo,
     expiryDate,
     expiryEpochNo,
     index,
-    title,
+    metadataStatus,
+    metadataValid,
     txHash,
     type,
-    isDataMissing,
+    title,
   } = proposal;
 
   const { isMobile, screenWidth } = useScreenDimension();
@@ -54,13 +55,13 @@ export const GovernanceVotedOnCard = ({ votedProposal, inProgress }: Props) => {
         justifyContent: "space-between",
         boxShadow: "0px 4px 15px 0px #DDE3F5",
         borderRadius: "20px",
-        backgroundColor: isDataMissing
+        backgroundColor: !metadataValid
           ? "rgba(251, 235, 235, 0.50)"
           : "rgba(255, 255, 255, 0.3)",
         // TODO: To decide if voted on cards can be actually in progress
         border: inProgress
           ? "1px solid #FFCBAD"
-          : isDataMissing
+          : !metadataValid
           ? "1px solid #F6D5D5"
           : "1px solid #C0E4BA",
       }}
@@ -76,11 +77,11 @@ export const GovernanceVotedOnCard = ({ votedProposal, inProgress }: Props) => {
       >
         <GovernanceActionCardHeader
           title={title}
-          isDataMissing={isDataMissing}
+          isDataMissing={metadataStatus}
         />
         <GovernanceActionCardElement
           label={t("govActions.abstract")}
-          text={about}
+          text={abstract}
           textVariant="twoLines"
           dataTestId="governance-action-abstract"
           isSliderCard