+++ govtool/frontend/src/components/organisms/DashboardTopNav.tsx
@@ -1,9 +1,10 @@
 import { useEffect, useState } from "react";
-import { Box, IconButton } from "@mui/material";
+import { Box, Button, IconButton } from "@mui/material";
+import { IconPlusCircle } from "@intersect.mbo/intersectmbo.org-icons-set";
 
 import { VotingPowerChips, Typography } from "@atoms";
-import { ICONS } from "@consts";
-import { useScreenDimension } from "@hooks";
+import { ICONS, PATHS, PDF_PATHS } from "@consts";
+import { useScreenDimension, useTranslation } from "@hooks";
 import { DashboardDrawerMobile } from "@organisms";
 
 type DashboardTopNavProps = {
@@ -17,12 +18,26 @@ export const DashboardTopNav = ({
   title,
   isVotingPowerHidden,
 }: DashboardTopNavProps) => {
+  const { t } = useTranslation();
+  const isProposalDiscussion = Object.values(PDF_PATHS).some(
+    (pdfPath) =>
+      window.location.pathname.includes(pdfPath) &&
+      window.location.pathname.includes(
+        PATHS.connectedProposalPillar.replace("/*", ""),
+      ),
+  );
   const [windowScroll, setWindowScroll] = useState<number>(0);
   const { isMobile } = useScreenDimension();
   const [isDrawerOpen, setIsDrawerOpen] = useState<boolean>(false);
 
   const openDrawer = () => {
     setIsDrawerOpen(true);
+  };
+
+  const goToProposalDiscussionCreateGovernanceAction = () => {
+    window.location.href = `${PATHS.connectedProposalPillar.replace("/*", "")}${
+      PDF_PATHS.proposalDiscussionCreateGovernanceAction
+    }`;
   };
 
   useEffect(() => {
@@ -85,6 +100,15 @@ export const DashboardTopNav = ({
             >
               <img alt="drawer" src={ICONS.drawerIcon} />
             </IconButton>
+          )}
+          {isProposalDiscussion && (
+            <Button
+              variant="contained"
+              startIcon={<IconPlusCircle fill="white" />}
+              onClick={goToProposalDiscussionCreateGovernanceAction}
+            >
+              {t("proposalDiscussion.proposeAGovernanceAction")}
+            </Button>
           )}
         </Box>
         {isMobile && (